extract getFileName function
