+++ src/index/generateTrees/toFractalTree.ts
@@ -9,18 +9,7 @@ import { isLinkedFile, linkedFileToOriginal } from "./shared/isLinkedFile";
 import chalk from "chalk";
 import { fileWithoutExtension } from "../shared/fileWithoutExtension";
 import { isTestFile } from "./shared/isTestFile";
-
-function getFileName(currentPath: string) {
-  const fileName = path.basename(currentPath, path.extname(currentPath));
-  const currentFolder = path.basename(path.dirname(currentPath));
-
-  if (fileName === "index" && currentFolder && currentFolder !== ".") {
-    return currentFolder;
-  }
-
-  return fileName;
-}
-
+import { getFileName } from "./toFractalTree/getFileName";
 interface FractalTreeConfig {
   nestMainModules: boolean;
 }
